Refactor GUI into submodules
